+++ Source/Boilerplate.Templates/Content/ApiTemplate/Startup.cs
@@ -78,7 +78,6 @@ namespace ApiTemplate
                 .AddMvcCore()
                     .AddApiExplorer()
                     .AddAuthorization()
-                    .AddFormatterMappings()
                     .AddDataAnnotations()
                     .AddJsonFormatters()
                     .AddCustomJsonOptions()